Split History into separate modules
