+++ ouroboros-consensus/src/Ouroboros/Consensus/HardFork/History/EpochInfo.hs
@@ -1,0 +1,54 @@
+{-# LANGUAGE FlexibleContexts    #-}
+{-# LANGUAGE RankNTypes          #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
+-- | Derive 'EpochInfo'
+module Ouroboros.Consensus.HardFork.History.EpochInfo (
+    summaryToEpochInfo
+  , snapshotEpochInfo
+  ) where
+
+import           Data.Functor.Identity
+
+import           Cardano.Slotting.EpochInfo.API
+
+import           Ouroboros.Consensus.Util.IOLike
+
+import           Ouroboros.Consensus.HardFork.History.Caching
+import           Ouroboros.Consensus.HardFork.History.Qry
+import           Ouroboros.Consensus.HardFork.History.Summary
+
+{-------------------------------------------------------------------------------
+  Translation to EpochInfo
+-------------------------------------------------------------------------------}
+
+-- | Construct 'EpochInfo' from a function that returns the hard fork summary
+--
+-- When a particular request fails with a 'PastHorizon' error, we ask for an
+-- updated summary, in the hope that the ledger state has advanced. If the query
+-- /still/ fails with that updated summary, the error is thrown as an exception.
+summaryToEpochInfo :: forall m xs. (MonadSTM m, MonadThrow (STM m))
+                   => STM m (Summary xs) -> m (EpochInfo (STM m))
+summaryToEpochInfo =
+    fmap go . runWithCachedSummary
+  where
+    go :: RunWithCachedSummary xs m -> EpochInfo (STM m)
+    go run = EpochInfo {
+          epochInfoSize_  = \e -> cachedRunQueryThrow run (QEpochSize   e)
+        , epochInfoFirst_ = \e -> cachedRunQueryThrow run (epochToSlot' e)
+        , epochInfoEpoch_ = \s -> cachedRunQueryThrow run (fst <$> slotToEpoch' s)
+        }
+
+-- | Construct an 'EpochInfo' for a /snapshot/ of the ledger state
+--
+-- When a particular request fails with a 'PastHorizon' error, we throw the
+-- error as a /pure/ exception. Such an exception would indicate a bug.
+snapshotEpochInfo :: forall xs. Summary xs -> EpochInfo Identity
+snapshotEpochInfo summary = EpochInfo {
+      epochInfoSize_  = \e -> runQueryPure' (QEpochSize   e)
+    , epochInfoFirst_ = \e -> runQueryPure' (epochToSlot' e)
+    , epochInfoEpoch_ = \s -> runQueryPure' (fst <$> slotToEpoch' s)
+    }
+  where
+    runQueryPure' :: Qry a -> Identity a
+    runQueryPure' = Identity . flip runQueryPure summary
